clean up pod encryption types in zk-sdk
